Update package config and docs for Docusaurus integration
Updated the package.json file to use Docusaurus for documentation. This will allow for an organized, interactive documentation setup for users. Several scripts were added to facilitate Docusaurus building, starting, serving, and deploying. Dependencies were also updated in favor of those used by Docusaurus, like React and Prism.

In addition to package.json, several markdown files were added to provide tutorials and welcome messages for the updated documentation site. Docs and blogs to cover various aspects for a site built with Docusaurus have been added.

README.md was also updated to reflect changes as regards to the new way of building, serving and deploying the site. In terms of version control, instructions have also been included as part of the documentations.

These changes are integral in making sure that the documentation is interactive, up to date and easy to navigate.